HCK-3427: separated indexes data manipulation logic
